+++ lib/store/sqlx/cached_conn.go
@@ -72,7 +72,7 @@ func (cc CachedConn) GetBit(key string, offset int64) (int, error) {
 	return cc.cache.GetBit(key, offset)
 }
 
-func (cc CachedConn) GetBits(key string, offset []uint) (map[uint]bool, error) {
+func (cc CachedConn) GetBits(key string, offset []int64) (map[int64]bool, error) {
 	return cc.cache.GetBits(key, offset)
 }
